requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
idna==2.9
pushbullet.py==0.11.0
python-magic==0.4.15
requests==2.23.0
six==1.14.0
surrogate==0.1
urllib3==1.25.8
w1thermsensor==1.3.0
websocket-client==0.57.0
freezegun~=0.3.15
